+++ tapo/src/requests/schedule.rs
@@ -1,0 +1,14 @@
+//! Request types for the plug "Schedule" feature.
+
+mod days_of_week;
+mod params;
+mod raw;
+mod schedule_rule;
+mod schedule_time;
+
+pub use days_of_week::*;
+pub use schedule_rule::*;
+pub use schedule_time::*;
+
+pub(crate) use params::*;
+pub(crate) use raw::*;