+++ tapo/src/requests/timer.rs
@@ -1,20 +1,20 @@
 use serde::Serialize;
 
-use crate::responses::TimerDesiredStateRaw;
+use crate::responses::{DesiredStateRaw, PowerState};
 
 #[derive(Debug, Clone, Serialize)]
 pub(crate) struct AddTimerParams {
     pub enable: bool,
     pub delay: u32,
-    pub desired_states: TimerDesiredStateRaw,
+    pub desired_states: DesiredStateRaw,
 }
 
 impl AddTimerParams {
-    pub(crate) fn new(delay_seconds: u32, on: bool) -> Self {
+    pub(crate) fn new(delay_seconds: u32, desired_state: PowerState) -> Self {
         Self {
             enable: true,
             delay: delay_seconds,
-            desired_states: TimerDesiredStateRaw { on },
+            desired_states: DesiredStateRaw::new(desired_state),
         }
     }
 }